deptcharge: Fix the return type of `timer_hz`
The return type of `timer_hz` is defined in libpayload.h. Since it is
changed to `uint32_t` in CL:5008569. We should also modify it in
deptcharge.

BRANCH=none
BUG=none
TEST=emerge-geralt libpayload depthcharge

Cq-Depend: chromium:5008569
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5010436

+++ src/base/timestamp.c
@@ -70,7 +70,7 @@ uint64_t get_us_since_boot(void)
 	if(!ts_table)
 		return timer_us(0);
 
-	uint64_t tick_freq_mhz = ts_table->tick_freq_mhz;
+	uint32_t tick_freq_mhz = ts_table->tick_freq_mhz;
 	if (!tick_freq_mhz)
 		tick_freq_mhz = timer_hz() / 1000000;
