+++ tests/lp_mocks/timer.c
@@ -10,7 +10,7 @@ uint64_t timer_raw_value(void)
 }
 
 __attribute__((weak))
-uint64_t timer_hz(void)
+uint32_t timer_hz(void)
 {
 	/* Libpayload requires it to be at least 1MHz.
 	   Keep it that way not to break anything */
